Added Latitude and Longitude information. Refactoring: longitude and latitude are now double values

+++ greenwatch-common/src/greenwatch/common/resource/PollutionResource.java
@@ -5,7 +5,7 @@ import greenwatch.common.vo.PollutionVO;
 import java.util.List;
 
 public interface PollutionResource {
-	List<PollutionVO> getPollutions(long lat, long lng);
+	List<PollutionVO> getPollutions(double lat, double lng);
 	void storePollution(PollutionVO pollution);
 	void updatePollution(PollutionVO pollution);
 }
